# common-platform-customer-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-customer-mgmt
  description: Customer Management Core Application
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.catalis.vc/customers
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/parties/{partyId}/legal-entities/{legalEntityId}:
    get:
      tags:
        - Legal Entities
      summary: Get legal entity by ID
      description: Retrieve a specific legal entity associated with a party by its unique identifier
      operationId: getLegalEntityById
      parameters:
        - name: partyId
          in: path
          description: Unique identifier of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: legalEntityId
          in: path
          description: Unique identifier of the legal entity
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Legal entity found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LegalEntityDTO'
        '404':
          description: Legal entity or party not found
    put:
      tags:
        - Legal Entities
      summary: Update legal entity
      description: Update an existing legal entity associated with a party
      operationId: updateLegalEntity
      parameters:
        - name: partyId
          in: path
          description: Unique identifier of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: legalEntityId
          in: path
          description: Unique identifier of the legal entity
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LegalEntityDTO'
        required: true
      responses:
        '200':
          description: Legal entity successfully updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LegalEntityDTO'
        '400':
          description: Invalid legal entity data
        '404':
          description: Legal entity or party not found
    delete:
      tags:
        - Legal Entities
      summary: Delete legal entity
      description: Delete a legal entity associated with a party from the system
      operationId: deleteLegalEntity
      parameters:
        - name: partyId
          in: path
          description: Unique identifier of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: legalEntityId
          in: path
          description: Unique identifier of the legal entity
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Legal entity successfully deleted
        '404':
          description: Legal entity or party not found
...
  /api/v1/parties/{partyId}/legal-entities:
    get:
      tags:
        - Legal Entities
      summary: Get legal entity by party ID
      description: Retrieve the legal entity associated with a specific party
      operationId: getLegalEntityByPartyId
      parameters:
        - name: partyId
          in: path
          description: Unique identifier of the party
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Legal entity found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LegalEntityDTO'
        '404':
          description: Party not found
    post:
      tags:
        - Legal Entities
      summary: Create legal entity for a party
      description: Create a new legal entity associated with a specific party
      operationId: createLegalEntity
      parameters:
        - name: partyId
          in: path
          description: Unique identifier of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LegalEntityDTO'
        required: true
      responses:
        '201':
          description: Legal entity successfully created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LegalEntityDTO'
        '400':
          description: Invalid legal entity data
        '404':
          description: Party not found
  /api/v1/parties/{partyId}/legal-entities/filter:
    post:
      tags:
        - Legal Entities
      summary: Filter legal entities for a party
      description: Retrieve a paginated list of legal entities associated with a specific party based on filtering criteria
      operationId: filterLegalEntities
      parameters:
        - name: partyId
          in: path
          description: Unique identifier of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestLegalEntityDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved filtered legal entities
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter request
        '404':
          description: Party not found
...
components:
  schemas:
...
    LegalEntityDTO:
      required:
        - legalName
        - partyId
      type: object
      properties:
        legalEntityId:
          type: string
          format: uuid
          readOnly: true
        partyId:
          type: string
          format: uuid
        legalName:
          maxLength: 200
          minLength: 0
          type: string
        tradeName:
          maxLength: 200
          minLength: 0
          type: string
        registrationNumber:
          maxLength: 100
          minLength: 0
          type: string
        taxIdNumber:
          maxLength: 50
          minLength: 0
          type: string
        legalFormId:
          type: string
          format: uuid
        incorporationDate:
          type: string
          format: date
        industryDescription:
          maxLength: 300
          minLength: 0
          type: string
        headcount:
          minimum: 0
          type: integer
          format: int64
        shareCapital:
          minimum: 0
          exclusiveMinimum: false
          type: number
        websiteUrl:
          maxLength: 500
          minLength: 0
          pattern: ^https?://.*
          type: string
        incorporationCountryId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    FilterRequestLegalEntityDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/LegalEntityDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Filter criteria for legal entities
...